pyproject.toml
[tool.poetry]
name = "network-maker"
version = "0.1.0"
description = ""
authors = ["a-ok123 <alexey@pastel.network>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
apscheduler = "^4.0.0a4"
pastel-gateway-sdk = "^0.1.1"
urllib3 = "^2.0.7"
diffusers = "^0.25.1"
transformers = "^4.37.0"
accelerate = "^0.26.1"
zipp = "^3.17.0"
huggingface-hub = "^0.20.2"
llama_cpp_python = "^0.2.33"
torch = {version = "^2.2.0+cu121", source = "pytorch"}
torchvision = {version = "^0.17.0+cu121", source = "pytorch"}
python-dotenv = "^1.0.1"
pydantic-settings = "^2.1.0"
compel = "^2.0.2"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
